+++ atdts/src/lib/Codegen.ml
@@ -34,9 +34,6 @@ let annot_schema : Atd.Annot.schema =
 let not_implemented loc msg =
   A.error_at loc ("not implemented in atdts: " ^ msg)
 
-let todo hint =
-  failwith ("TODO: " ^ hint)
-
 (* This is used to convince the TypeScript type checker in strict mode
    that a function doesn't lack a return statement (TS2366) *)
 let impossible = "throw new Error('impossible')"
@@ -90,10 +87,12 @@ let spaced ?(spacer = [Line ""]) (blocks : B.node list) : B.node list =
 (*
    Eliminate the 'wrap' constructs since we don't do anything with them,
    and produce decent error messages for the unsupported constructs.
+   The result is guaranteed to not be of the form 'Wrap ...' but it may
+   contain 'Wrap' constructs within its type arguments.
 *)
-let rec unwrap e =
+let rec shallow_unwrap e =
   match e with
-  | Wrap (loc, e, an) -> unwrap e
+  | Wrap (loc, e, an) -> shallow_unwrap e
   | Shared (loc, e, an) -> not_implemented loc "cyclic references"
   | Tvar _
   | Sum _
@@ -627,7 +626,7 @@ let rec type_name_of_expr env (e : type_expr) : string =
   | Option (loc, e, an) -> sprintf "Option<%s>" (type_name_of_expr env e)
   | Nullable (loc, e, an) -> sprintf "(%s | null)" (type_name_of_expr env e)
   | Shared (loc, e, an) -> not_implemented loc "shared"
-  | Wrap (loc, e, an) -> todo "wrap"
+  | Wrap (loc, e, an) -> type_name_of_expr env e
   | Name (loc, (loc2, name, []), an) -> ts_type_name env name
   | Name (loc, _, _) -> assert false
   | Tvar (loc, _) -> not_implemented loc "type variables"
@@ -861,7 +860,7 @@ let sum_type env loc name cases =
 let make_type_def env ((loc, (name, param, an), e) : A.type_def) : B.t =
   if param <> [] then
     not_implemented loc "parametrized type";
-  match unwrap e with
+  match shallow_unwrap e with
   | Sum (loc, variants, an) ->
       sum_type env loc name (flatten_variants variants)
   | Record (loc, fields, an) ->
@@ -916,7 +915,7 @@ let write_case env loc orig_name an opt_e =
       ]
 
 let read_root_expr env ~ts_type_name e =
-  match unwrap e with
+  match shallow_unwrap e with
   | Sum (loc, variants, an) ->
       let cases = flatten_variants variants in
       let cases0, cases1 =
@@ -1042,7 +1041,7 @@ let read_root_expr env ~ts_type_name e =
   | Tvar _ -> assert false
 
 let write_root_expr env ~ts_type_name e =
-  match unwrap e with
+  match shallow_unwrap e with
   | Sum (loc, variants, an) ->
       let cases = flatten_variants variants in
       [